moving control out of main namespace
